trading_bot/requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
